+++ src/test/java/wooteco/subway/dao/InmemoryStationDaoTest.java
@@ -1,8 +1,6 @@
 package wooteco.subway.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -40,34 +38,9 @@ class InmemoryStationDaoTest {
     @DisplayName("Station을 삭제할 수 있다.")
     void delete() {
         Station station = inmemoryStationDao.save(new Station("오리"));
-        Long stationId = station.getId();
-
-        assertThatCode(() -> inmemoryStationDao.delete(stationId)).doesNotThrowAnyException();
+        inmemoryStationDao.delete(station.getId());
+        assertThat(inmemoryStationDao.existById(station.getId())).isFalse();
     }
-
-    @Test
-    @DisplayName("없는 id의 Station을 삭제할 수 없다.")
-    void deleteByInvalidId() {
-        Station station = inmemoryStationDao.save(new Station("오리"));
-        Long stationId = station.getId() + 1;
-
-        assertThatThrownBy(() -> inmemoryStationDao.delete(stationId))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("없는 station 입니다.");
-    }
-
-    @Test
-    @DisplayName("이미 삭제한 id의 Station을 삭제할 수 없다.")
-    void deleteByDuplicatedId() {
-        Station station = inmemoryStationDao.save(new Station("오리"));
-        Long stationId = station.getId();
-        inmemoryStationDao.delete(stationId);
-
-        assertThatThrownBy(() -> inmemoryStationDao.delete(stationId))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("없는 station 입니다.");
-    }
-
 
     @Test
     @DisplayName("Station 이름이 존재하는지 확인할 수 있다.")
