+++ src/test/java/wooteco/subway/dao/JdbcStationDaoTest.java
@@ -1,8 +1,6 @@
 package wooteco.subway.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -65,31 +63,7 @@ class JdbcStationDaoTest {
     @DisplayName("Station을 삭제할 수 있다.")
     void delete() {
         Station station = stationDao.save(new Station("오리"));
-        Long stationId = station.getId();
-
-        assertThatCode(() -> stationDao.delete(stationId)).doesNotThrowAnyException();
-    }
-
-    @Test
-    @DisplayName("없는 id의 Station을 삭제할 수 없다.")
-    void deleteByInvalidId() {
-        Station station = stationDao.save(new Station("오리"));
-        Long stationId = station.getId() + 1;
-
-        assertThatThrownBy(() -> stationDao.delete(stationId))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("없는 station 입니다.");
-    }
-
-    @Test
-    @DisplayName("이미 삭제한 id의 Station을 삭제할 수 없다.")
-    void deleteByDuplicatedId() {
-        Station station = stationDao.save(new Station("오리"));
-        Long stationId = station.getId();
-        stationDao.delete(stationId);
-
-        assertThatThrownBy(() -> stationDao.delete(stationId))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("없는 station 입니다.");
+        stationDao.delete(station.getId());
+        assertThat(stationDao.existById(station.getId())).isFalse();
     }
 }
