+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -2,18 +2,21 @@ package wooteco.subway.service;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.util.NoSuchElementException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
 import wooteco.subway.domain.Station;
 
 @SpringBootTest
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 class StationServiceTest {
 
     @Autowired
     private StationService stationService;
-    
+
     @Test
     @DisplayName("이미 존재하는 역 이름이 있을 때 예외가 발생한다.")
     void saveExceptionByDuplicatedName() {
@@ -21,5 +24,28 @@ class StationServiceTest {
         assertThatThrownBy(() -> stationService.save(new Station("오리")))
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessage("이미 존재하는 역 이름입니다.");
+    }
+
+    @Test
+    @DisplayName("없는 id의 Station을 삭제할 수 없다.")
+    void deleteByInvalidId() {
+        Station station = stationService.save(new Station("오리"));
+        Long stationId = station.getId() + 1;
+
+        assertThatThrownBy(() -> stationService.delete(stationId))
+                .isInstanceOf(NoSuchElementException.class)
+                .hasMessage("없는 Station 입니다.");
+    }
+
+    @Test
+    @DisplayName("이미 삭제한 id의 Station을 삭제할 수 없다.")
+    void deleteByDuplicatedId() {
+        Station station = stationService.save(new Station("오리"));
+        Long stationId = station.getId();
+        stationService.delete(stationId);
+
+        assertThatThrownBy(() -> stationService.delete(stationId))
+                .isInstanceOf(NoSuchElementException.class)
+                .hasMessage("없는 Station 입니다.");
     }
 }
